refactor reducer and tests
