+++ src/actionType.js
@@ -1,0 +1,8 @@
+export default {
+  SET_CURRENT_CITY: `SET_CURRENT_CITY`,
+  GET_OFFERS_LIST: `GET_OFFERS_LIST`,
+  SET_ACTIVE_OFFER: `SET_ACTIVE_OFFER`,
+  SET_FOCUS_OFFER: `SET_FOCUS_OFFER`,
+  REMOVE_ACTIVE_OFFER: `REMOVE_ACTIVE_OFFER`,
+  GET_AVAILABLE_OFFERS: `GET_AVAILABLE_OFFERS`,
+};
