+++ src/components/place-card/place-card.test.js
@@ -5,47 +5,30 @@ import configureStore from "redux-mock-store";
 import {
   Provider
 } from "react-redux";
+import initialState from "../../mocks/initialState.js";
 
 const mockStore = configureStore([]);
-
 const handleHeaderOfferClick = () => {};
 const handlePlaceCardMouseEnter = () => {};
 const handlePlaceCardMouseLeave = () => {};
+const offer = {
+  src: [
+    `img/apartment-03.jpg`,
+    `img/studio-01.jpg`,
+    `img/apartment-01.jpg`,
+    `img/room.jpg`,
+    `img/apartment-01.jpg`,
+    `img/apartment-02.jpg`,
+  ],
+  price: 120,
+  rating: 4,
+  name: `Beautiful &amp; luxurious apartment at great location`,
+  type: `Apartment`,
+  mark: true,
+};
 
 it(`Render place card`, () => {
-  const store = mockStore({
-    offer: {
-      id: 123,
-      src: [
-        `img/apartment-03.jpg`,
-        `img/studio-01.jpg`,
-        `img/apartment-01.jpg`,
-        `img/room.jpg`,
-        `img/apartment-01.jpg`,
-        `img/apartment-02.jpg`,
-      ],
-      price: 120,
-      rating: 4,
-      name: `Beautiful &amp; luxurious apartment at great location`,
-      description: `An independent House, strategically located between Rembrand Square and National Opera,`
-      + ` but where the bustle of the city comes to rest in this alley flowery and colorful.`,
-      type: `Apartment`,
-      mark: true,
-      insideItems: [
-        `Wi-Fi`,
-        `Washing machine`,
-        `Towels`,
-        `Heating`,
-        `Coffee machine`,
-        `Baby seat`,
-        `Kitchen`,
-        `Dishwasher`,
-        `Cabel TV`,
-        `Fridge`
-      ],
-      coordinate: [52.3909553943508, 4.85309666406198],
-    }
-  });
+  const store = mockStore(initialState);
   const tree = renderer
     .create(
         <Provider store={store}>
@@ -58,6 +41,9 @@ it(`Render place card`, () => {
             }
             onMouseLeave={
               handlePlaceCardMouseLeave
+            }
+            offer={
+              offer
             }
           />
         </Provider>
